:truck: move: templateディレクトリを新設し、移動
